// services/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
};

export type BlockChangedFilter = {
  number_gte: Scalars['Int'];
};

export type Block_Height = {
  hash?: InputMaybe<Scalars['Bytes']>;
  number?: InputMaybe<Scalars['Int']>;
  number_gte?: InputMaybe<Scalars['Int']>;
};

export type HistoryEvent = {
  __typename?: 'HistoryEvent';
  amount?: Maybe<Scalars['BigInt']>;
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  eventType: Scalars['String'];
  id: Scalars['ID'];
  oldStream?: Maybe<Stream>;
  stream?: Maybe<Stream>;
  token: Token;
  txHash: Scalars['Bytes'];
  users: Array<User>;
};
// ...
export type LlamaPayContract = {
  __typename?: 'LlamaPayContract';
  address: Scalars['Bytes'];
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  factory: LlamaPayFactory;
  id: Scalars['ID'];
  streams: Array<Stream>;
  token: Token;
};
// ...
export type LlamaPayFactory = {
  __typename?: 'LlamaPayFactory';
  address: Scalars['Bytes'];
  contracts: Array<LlamaPayContract>;
  count: Scalars['Int'];
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  id: Scalars['ID'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  historyEvent?: Maybe<HistoryEvent>;
  historyEvents: Array<HistoryEvent>;
  llamaPayContract?: Maybe<LlamaPayContract>;
  llamaPayContracts: Array<LlamaPayContract>;
  llamaPayFactories: Array<LlamaPayFactory>;
  llamaPayFactory?: Maybe<LlamaPayFactory>;
  stream?: Maybe<Stream>;
  streams: Array<Stream>;
  token?: Maybe<Token>;
  tokens: Array<Token>;
  user?: Maybe<User>;
  users: Array<User>;
};
// ...
export type Stream = {
  __typename?: 'Stream';
  active: Scalars['Boolean'];
  amountPerSec: Scalars['BigInt'];
  contract: LlamaPayContract;
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  historicalEvents: Array<HistoryEvent>;
  id: Scalars['ID'];
  lastPaused: Scalars['BigInt'];
  paused: Scalars['Boolean'];
  pausedAmount: Scalars['BigInt'];
  payee: User;
  payer: User;
  reason?: Maybe<Scalars['String']>;
  streamId: Scalars['Bytes'];
  token: Token;
  users: Array<User>;
};
// ...
export type Token = {
  __typename?: 'Token';
  address: Scalars['Bytes'];
  contract: LlamaPayContract;
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  decimals: Scalars['Int'];
  id: Scalars['ID'];
  name: Scalars['String'];
  symbol: Scalars['String'];
};
// ...
export type User = {
  __typename?: 'User';
  address: Scalars['Bytes'];
  createdBlock: Scalars['BigInt'];
  createdTimestamp: Scalars['BigInt'];
  historicalEvents: Array<HistoryEvent>;
  id: Scalars['ID'];
  streams: Array<Stream>;
};
// ...
export type _Meta_ = {
  __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars['String'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars['Boolean'];
};
// ...
export type StreamAndHistoryQuery = { __typename?: 'Query', user?: { __typename?: 'User', streams: Array<{ __typename?: 'Stream', streamId: any, active: boolean, reason?: string | null, paused: boolean, pausedAmount: any, lastPaused: any, amountPerSec: any, createdTimestamp: any, contract: { __typename?: 'LlamaPayContract', address: any }, payer: { __typename?: 'User', id: string }, payee: { __typename?: 'User', id: string }, token: { __typename?: 'Token', address: any, name: string, decimals: number, symbol: string }, historicalEvents: Array<{ __typename?: 'HistoryEvent', eventType: string, txHash: any, createdTimestamp: any }> }>, historicalEvents: Array<{ __typename?: 'HistoryEvent', txHash: any, eventType: string, amount?: any | null, createdTimestamp: any, users: Array<{ __typename?: 'User', id: string }>, stream?: { __typename?: 'Stream', streamId: any, amountPerSec: any, createdTimestamp: any, payer: { __typename?: 'User', id: string }, payee: { __typename?: 'User', id: string }, token: { __typename?: 'Token', address: any, name: string, decimals: number, symbol: string } } | null, oldStream?: { __typename?: 'Stream', streamId: any, amountPerSec: any, createdTimestamp: any, payer: { __typename?: 'User', id: string }, payee: { __typename?: 'User', id: string }, token: { __typename?: 'Token', address: any, symbol: string } } | null, token: { __typename?: 'Token', symbol: string, decimals: number } }> } | null };